Fixed report select splicing wrong element
The report select script was finding the index with a lamda but it always just returned the wrong index which left that target game in the list and an other game would then be reported.
this was a game-breaking bug as it left some people in an unreported game and you had to restart the bot to fix this.